ContextPropagatingThreadFactory: move to o.b.core package
This is a step towards deprecating `Context` entirely. The move will
allow ContextPropagatingThreadFactory to use (soon-to-be) package private
methods in `Context` and will also allow ContextPropagatingThreadFactory
itself to be made package-private.
