+++ core/src/main/java/org/bitcoinj/core/ContextPropagatingThreadFactory.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.utils;
+package org.bitcoinj.core;
 
-import org.bitcoinj.core.Context;
+import org.bitcoinj.utils.Threading;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
